+++ CHANGELOG.md
@@ -16,6 +16,11 @@
 
 ---
 
+## v3.6.3
+
+- [REPLAY-328][rumf-1035] gracefully handle recorder Worker initialisation crash ([#1068](https://github.com/DataDog/browser-sdk/pull/1068))
+- 🔊 Add monitoring on session type change ([#1075](https://github.com/DataDog/browser-sdk/pull/1075))
+
 ## v3.6.2
 
 - 🔊 Add extra monitoring info on lite session with replay ([#1072](https://github.com/DataDog/browser-sdk/pull/1072))